+++ CHANGELOG.md
@@ -60,6 +60,7 @@ This is a major update with several breaking changes.
   - `Lumberjack::Logger#tag_globally`
   - `Lumberjack::Logger#remove_tag`
   - `Lumberjack::LogEntry#tag`
+  - `Lumberjack::LogEntry#tags`
   - `Lumberjack::LogEntry#nested_tags`
   - `Lumberjack::Logger::Utils.flatten_tags`
   - `Lumberjack::Logger::Utils.expand_tags`